Fuck you, Fuck code i did this shit in the most scuffed way as possible
