+++ main/website/test.py
@@ -1,0 +1,8 @@
+api = None
+
+def SetApi(setApi):
+    global api
+    api = setApi
+
+def returnApi():
+    return api
