+++ server/logs/latest.log
@@ -1,0 +1,51 @@
+[17:17:54] [main/INFO]: Environment: authHost='https://authserver.mojang.com', accountsHost='https://api.mojang.com', sessionHost='https://sessionserver.mojang.com', servicesHost='https://api.minecraftservices.com', name='PROD'
+[17:17:55] [main/WARN]: Ambiguity between arguments [teleport, destination] and [teleport, targets] with inputs: [Player, 0123, @e, dd12be42-52a9-4a91-a8a1-11c01849e498]
+[17:17:55] [main/WARN]: Ambiguity between arguments [teleport, location] and [teleport, destination] with inputs: [0.1 -0.5 .9, 0 0 0]
+[17:17:55] [main/WARN]: Ambiguity between arguments [teleport, location] and [teleport, targets] with inputs: [0.1 -0.5 .9, 0 0 0]
+[17:17:55] [main/WARN]: Ambiguity between arguments [teleport, targets] and [teleport, destination] with inputs: [Player, 0123, dd12be42-52a9-4a91-a8a1-11c01849e498]
+[17:17:55] [main/WARN]: Ambiguity between arguments [teleport, targets, location] and [teleport, targets, destination] with inputs: [0.1 -0.5 .9, 0 0 0]
+[17:17:55] [main/INFO]: Reloading ResourceManager: Default
+[17:17:56] [Worker-Main-12/INFO]: Loaded 7 recipes
+[17:17:56] [Worker-Main-12/INFO]: Loaded 927 advancements
+[17:17:58] [Server thread/INFO]: Starting minecraft server version 1.16.4
+[17:17:58] [Server thread/INFO]: Loading properties
+[17:17:58] [Server thread/INFO]: Default game type: SURVIVAL
+[17:17:58] [Server thread/INFO]: Generating keypair
+[17:17:58] [Server thread/INFO]: Starting Minecraft server on *:25565
+[17:17:58] [Server thread/INFO]: Using default channel type
+[17:17:58] [Server thread/INFO]: Preparing level "world"
+[17:17:58] [Server thread/INFO]: Preparing start region for dimension minecraft:overworld
+[17:18:00] [Server thread/INFO]: Preparing spawn area: 0%
+[17:18:00] [Server thread/INFO]: Preparing spawn area: 0%
+[17:18:00] [Server thread/INFO]: Preparing spawn area: 0%
+[17:18:00] [Server thread/INFO]: Preparing spawn area: 0%
+[17:18:00] [Server thread/INFO]: Preparing spawn area: 0%
+[17:18:04] [Worker-Main-14/INFO]: Preparing spawn area: 83%
+[17:18:04] [Worker-Main-14/INFO]: Preparing spawn area: 83%
+[17:18:04] [Worker-Main-9/INFO]: Preparing spawn area: 83%
+[17:18:04] [Worker-Main-12/INFO]: Preparing spawn area: 83%
+[17:18:04] [Worker-Main-9/INFO]: Preparing spawn area: 83%
+[17:18:04] [Worker-Main-9/INFO]: Preparing spawn area: 83%
+[17:18:04] [Worker-Main-12/INFO]: Preparing spawn area: 83%
+[17:18:04] [Worker-Main-9/INFO]: Preparing spawn area: 87%
+[17:18:05] [Worker-Main-12/INFO]: Preparing spawn area: 87%
+[17:18:05] [Worker-Main-9/INFO]: Preparing spawn area: 87%
+[17:18:06] [Worker-Main-12/INFO]: Preparing spawn area: 87%
+[17:18:06] [Worker-Main-14/INFO]: Preparing spawn area: 88%
+[17:18:07] [Worker-Main-9/INFO]: Preparing spawn area: 89%
+[17:18:07] [Server thread/INFO]: Preparing spawn area: 90%
+[17:18:08] [Server thread/INFO]: Time elapsed: 9253 ms
+[17:18:08] [Server thread/INFO]: Done (9.365s)! For help, type "help"
+[17:18:10] [Server thread/INFO]: Stopping the server
+[17:18:10] [Server thread/INFO]: Stopping server
+[17:18:10] [Server thread/INFO]: Saving players
+[17:18:10] [Server thread/INFO]: Saving worlds
+[17:18:10] [Server thread/INFO]: Saving chunks for level 'ServerLevel[world]'/minecraft:overworld
+[17:18:11] [Server thread/INFO]: ThreadedAnvilChunkStorage (world): All chunks are saved
+[17:18:11] [Server thread/INFO]: Saving chunks for level 'ServerLevel[world]'/minecraft:the_nether
+[17:18:11] [Server thread/INFO]: ThreadedAnvilChunkStorage (DIM-1): All chunks are saved
+[17:18:11] [Server thread/INFO]: Saving chunks for level 'ServerLevel[world]'/minecraft:the_end
+[17:18:11] [Server thread/INFO]: ThreadedAnvilChunkStorage (DIM1): All chunks are saved
+[17:18:11] [Server thread/INFO]: ThreadedAnvilChunkStorage (world): All chunks are saved
+[17:18:11] [Server thread/INFO]: ThreadedAnvilChunkStorage (DIM-1): All chunks are saved
+[17:18:11] [Server thread/INFO]: ThreadedAnvilChunkStorage (DIM1): All chunks are saved
