+++ server/server.properties
@@ -1,5 +1,5 @@
 #Minecraft server properties
-#Fri Jan 08 15:26:54 EST 2021
+#Fri Jan 08 17:17:54 EST 2021
 enable-jmx-monitoring=false
 rcon.port=25575
 level-seed=
@@ -15,8 +15,8 @@ generate-structures=true
 difficulty=easy
 network-compression-threshold=256
 max-tick-time=60000
-max-players=20
 use-native-transport=true
+max-players=20
 online-mode=true
 enable-status=true
 allow-flight=false
